Added ability to modify total thread count
I added an additional argument to the command line to allow max threads to be defined when running the scan. If left blank, it will run with the max available to ThreadPoolExecutor being defined by the length of the list of ports (total ports to be scanned equals max available threads), however now if '-t' is defined as 10, the max available argument for the ThreadPoolExecutor function will be limited to 10 (or any number chosen by the user). This allows the user to decide to use less to make a scan less aggressive, or to potentially limit a spike in memory usage if using this program on a machine with limited resources.